Normalisation by introduction of DigaVe
This is a follow-up of the conference call this Wednesday. Maybe it's too late altogether for a change - I could live it. 

Regarding the actual modification: A Diga itself can have several different DigaVe-s each with only a different DigaVeId, PZN, AnwendungsTage, Packungspreis. All the rest is the same (to be discussed). Thus the previous structure IMHO would be 1) redundant 2) possible error prone 3) more cumbersome for later storing in a database to distribute.

The edit here is just for demonstration, the XSD would have to be modified of course as well.

The sub-elements of DigaVe could be modeled as attributes which possibly would make it slightly easier to read, likewise DigaId could be an attribute of Diga.